+++ layers/Engine/packages/component-model/src/DirectiveResolvers/AbstractFieldDirectiveResolver.php
@@ -1411,6 +1411,16 @@ abstract class AbstractFieldDirectiveResolver extends AbstractDirectiveResolver 
         ObjectResolutionFeedbackStore|SchemaFeedbackStore $schemaOrObjectResolutionFeedbackStore,
     ): void {
         /**
+         * After setting fields as null below, the references to the fields inside $idFieldSetToRemove
+         * will also be removed.
+         *
+         * Hence, make a duplicate first, to pass to the feedback store.
+         */
+        $idFieldSetForFeedback = [];
+        foreach ($idFieldSetToRemove as $id => $fieldSet) {
+            $idFieldSetForFeedback[$id] = new EngineIterationFieldSet($fieldSet->fields);
+        }
+        /**
          * Remove the fields from the directive pipeline
          */
         /** @var ModuleConfiguration */
@@ -1434,7 +1444,7 @@ abstract class AbstractFieldDirectiveResolver extends AbstractDirectiveResolver 
                     $feedbackItemResolution,
                     $astNode,
                     $relationalTypeResolver,
-                    MethodHelpers::getFieldsFromIDFieldSet($idFieldSetToRemove)
+                    MethodHelpers::getFieldsFromIDFieldSet($idFieldSetForFeedback)
                 )
             );
             return;
@@ -1448,7 +1458,7 @@ abstract class AbstractFieldDirectiveResolver extends AbstractDirectiveResolver 
                 $astNode,
                 $relationalTypeResolver,
                 $this->directive,
-                $idFieldSetToRemove
+                $idFieldSetForFeedback
             )
         );
     }