+++ layers/Engine/packages/component-model/src/TypeResolvers/InputObjectType/AbstractQueryableInputObjectTypeResolver.php
@@ -109,7 +109,7 @@ abstract class AbstractQueryableInputObjectTypeResolver extends AbstractInputObj
         /**
          * If the input field is an InputObject, recursively apply this function
          */
-        if ($isQueryableInputObjectTypeResolver) {
+        if ($isQueryableInputObjectTypeResolver && $inputFieldValue !== null) {
             /** @var QueryableInputObjectTypeResolverInterface $queryableInputObjectTypeResolver */
             $queryableInputObjectTypeResolver->integrateInputValueToFilteringQueryArgs($query, $inputFieldValue);
         }