+++ layers/WPSchema/packages/blocks/src/FieldResolvers/ObjectType/CustomPostObjectTypeFieldResolver.php
@@ -15,6 +15,7 @@ use PoP\ComponentModel\Schema\SchemaTypeModifiers;
 use PoP\ComponentModel\TypeResolvers\ConcreteTypeResolverInterface;
 use PoP\ComponentModel\TypeResolvers\InputTypeResolverInterface;
 use PoP\ComponentModel\TypeResolvers\ObjectType\ObjectTypeResolverInterface;
+use PoP\ComponentModel\TypeResolvers\ScalarType\BooleanScalarTypeResolver;
 use PoP\Engine\TypeResolvers\ScalarType\JSONObjectScalarTypeResolver;
 use PoP\GraphQLParser\Spec\Parser\Ast\FieldInterface;
 use PoPCMSSchema\CustomPosts\TypeResolvers\ObjectType\AbstractCustomPostObjectTypeResolver;
@@ -25,6 +26,7 @@ use PoPWPSchema\Blocks\Constants\HookNames;
 use PoPWPSchema\Blocks\ObjectModels\BlockInterface;
 use PoPWPSchema\Blocks\ObjectModels\GeneralBlock;
 use PoPWPSchema\Blocks\TypeHelpers\BlockUnionTypeHelpers;
+use PoPWPSchema\Blocks\TypeResolvers\InputObjectType\IncludeBlockPropertiesInputObjectTypeResolver;
 use stdClass;
 use WP_Post;
 
@@ -33,6 +35,8 @@ class CustomPostObjectTypeFieldResolver extends AbstractQueryableObjectTypeField
     private ?BlockContentParserInterface $blockContentParser = null;
     private ?JSONObjectScalarTypeResolver $jsonObjectScalarTypeResolver = null;
     private ?IncludeExcludeFilterInputObjectTypeResolver $includeExcludeFilterInputObjectTypeResolver = null;
+    private ?IncludeBlockPropertiesInputObjectTypeResolver $includeBlockPropertiesInputObjectTypeResolver = null;
+    private ?BooleanScalarTypeResolver $booleanScalarTypeResolver = null;
 
     final protected function getBlockContentParser(): BlockContentParserInterface
     {
@@ -60,6 +64,24 @@ class CustomPostObjectTypeFieldResolver extends AbstractQueryableObjectTypeField
             $this->includeExcludeFilterInputObjectTypeResolver = $includeExcludeFilterInputObjectTypeResolver;
         }
         return $this->includeExcludeFilterInputObjectTypeResolver;
+    }
+    final protected function getIncludeBlockPropertiesInputObjectTypeResolver(): IncludeBlockPropertiesInputObjectTypeResolver
+    {
+        if ($this->includeBlockPropertiesInputObjectTypeResolver === null) {
+            /** @var IncludeBlockPropertiesInputObjectTypeResolver */
+            $includeBlockPropertiesInputObjectTypeResolver = $this->instanceManager->getInstance(IncludeBlockPropertiesInputObjectTypeResolver::class);
+            $this->includeBlockPropertiesInputObjectTypeResolver = $includeBlockPropertiesInputObjectTypeResolver;
+        }
+        return $this->includeBlockPropertiesInputObjectTypeResolver;
+    }
+    final protected function getBooleanScalarTypeResolver(): BooleanScalarTypeResolver
+    {
+        if ($this->booleanScalarTypeResolver === null) {
+            /** @var BooleanScalarTypeResolver */
+            $booleanScalarTypeResolver = $this->instanceManager->getInstance(BooleanScalarTypeResolver::class);
+            $this->booleanScalarTypeResolver = $booleanScalarTypeResolver;
+        }
+        return $this->booleanScalarTypeResolver;
     }
 
     /**
@@ -126,17 +148,42 @@ class CustomPostObjectTypeFieldResolver extends AbstractQueryableObjectTypeField
     {
         $fieldArgNameTypeResolvers = parent::getFieldArgNameTypeResolvers($objectTypeResolver, $fieldName);
         return match ($fieldName) {
-            'blocks',
+            'blocks'
+                => array_merge(
+                    $fieldArgNameTypeResolvers,
+                    [
+                        'filterBy' => $this->getIncludeExcludeFilterInputObjectTypeResolver(),
+                        'useHTML5Parser' => $this->getBooleanScalarTypeResolver(),
+                    ]
+                ),
             'blockDataItems',
             'blockFlattenedDataItems'
                 => array_merge(
                     $fieldArgNameTypeResolvers,
                     [
                         'filterBy' => $this->getIncludeExcludeFilterInputObjectTypeResolver(),
+                        'includeProperties' => $this->getIncludeBlockPropertiesInputObjectTypeResolver(),
+                        'useHTML5Parser' => $this->getBooleanScalarTypeResolver(),
                     ]
                 ),
             default
                 => $fieldArgNameTypeResolvers,
+        };
+    }
+
+    public function getFieldArgDescription(ObjectTypeResolverInterface $objectTypeResolver, string $fieldName, string $fieldArgName): ?string
+    {
+        return match ($fieldArgName) {
+            'useHTML5Parser' => $this->__('Indicate if to use the HTML5 parser when converting block content. When parsing HTML5, the parser will convert non-regular breaking spaces, such as "U+00A0", to "&nbsp;" entities.', 'blocks'),
+            default => parent::getFieldArgDescription($objectTypeResolver, $fieldName, $fieldArgName),
+        };
+    }
+
+    public function getFieldArgDefaultValue(ObjectTypeResolverInterface $objectTypeResolver, string $fieldName, string $fieldArgName): mixed
+    {
+        return match ($fieldArgName) {
+            'useHTML5Parser' => true,
+            default => parent::getFieldArgDefaultValue($objectTypeResolver, $fieldName, $fieldArgName),
         };
     }
 
@@ -156,7 +203,28 @@ class CustomPostObjectTypeFieldResolver extends AbstractQueryableObjectTypeField
                 /** @var stdClass|null */
                 $filterBy = $fieldDataAccessor->getValue('filterBy');
 
-                $options = [];
+                $includeInnerContent = false;
+                if ($fieldName === 'blocks') {
+                    $includeInnerContent = true;
+                } else {
+                    /** @var stdClass|null */
+                    $includeProperties = $fieldDataAccessor->getValue('includeProperties');
+                    $includeInnerContent = $includeProperties->innerContent ?? false;
+                }
+
+                /**
+                 * When parsing HTML5, the parser will convert non-breaking space "U+00A0" to "&nbsp;" entities,
+                 * and then the search/replace in "rawContent" will fail.
+                 * Then allow to disable the HTML5 parser.
+                 *
+                 * @var bool
+                 */
+                $useHTML5Parser = $fieldDataAccessor->getValue('useHTML5Parser') ?? true;
+
+                $options = [
+                    'include-inner-content' => $includeInnerContent,
+                    'use-html5-parser' => $useHTML5Parser,
+                ];
 
                 /**
                  * Add the filtering options.
@@ -186,10 +254,6 @@ class CustomPostObjectTypeFieldResolver extends AbstractQueryableObjectTypeField
                         $filterOptions['exclude'] = $filterBy->exclude;
                     }
                     $options['filter'] = $filterOptions;
-
-                    if ($fieldName === 'blocks') {
-                        $options['include-inner-content'] = true;
-                    }
                 }
 
                 $blockContentParserPayload = null;