+++ layers/GatoGraphQLForWP/plugins/gatographql/CHANGELOG.md
@@ -4,11 +4,29 @@ All notable changes to `gatographql` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
-## 15.1.1 - DATE
+## 15.2.0 - 26/11/2025
+
+### Added
+
+- Added field `_urlToCustomPostID` to get the custom post ID from a URL (#3229)
+- Support `--fail-if-log-notifications` param to WP-CLI commands, to print IDs only when a log entry was added during the execution (#3221)
+- Documentation for new field `_htmlParseHTML5` in the Helper Function Collection extension (#3237)
+
+### Improvements
+
+- Include block's `innerContent` property for `blockDataItems` and `blockFlattenedDataItems` fields (#3224)
+- `CustomPost.slug` can return empty string if not yet set (#3228)
+- In fields `blocks`, `blockDataItems`, and `blockFlattenedDataItems`, allow to not convert block content using HTML5 parser (#3230)
+- Tested up to WordPress 6.9 (#3238)
 
 ### Fixed
 
 - Check slugPath is not empty when adding page parent by slugPath (#3218)
+- Exception when passing field arg with value null (#3222)
+- Querying taxonomy that is assigned to more than 1 CPT (#3225)
+- Errors inside nested directives not passing the fields upwards when `nestErrorsInMetaDirectives` is false (#3226)
+- The link on the `wpAdminEditURL` field to print `&` as `&` (not `&amp`) (#3227)
+- "Undefined array key" warning (#3239)
 
 ## 15.1.0 - 17/10/2025